Extract munit diff module
