+++ munit/shared/src/main/scala/munit/ComparisonFailExceptionHandler.scala
@@ -1,0 +1,10 @@
+package munit
+
+trait ComparisonFailExceptionHandler {
+  def handle(
+      message: String,
+      obtained: String,
+      expected: String,
+      location: Location
+  ): Nothing
+}
