+++ scala-diff/shared/src/main/scala/munit/diff/Delta.scala
@@ -1,4 +1,4 @@
-package munit.internal.difflib
+package munit.diff
 
 sealed abstract class Delta[T](original: Chunk[T], revised: Chunk[T]) {
