+++ scala-diff/shared/src/main/scala/munit/diff/DifferentiationFailedException.scala
@@ -1,3 +1,3 @@
-package munit.internal.difflib
+package munit.diff
 
 class DifferentiationFailedException(message: String) extends Exception(message)
