+++ scala-diff/shared/src/main/scala/munit/diff/Diffs.scala
@@ -1,0 +1,28 @@
+package munit.diff
+
+object Diffs {
+
+  def create(obtained: String, expected: String): Diff =
+    new Diff(obtained, expected)
+
+  def createDiffOnlyReport(
+      obtained: String,
+      expected: String
+  ): String = {
+    create(obtained, expected).createDiffOnlyReport()
+  }
+
+  def createReport(
+      obtained: String,
+      expected: String,
+      title: String,
+      printObtainedAsStripMargin: Boolean = true
+  ): String = {
+    create(obtained, expected).createReport(title, printObtainedAsStripMargin)
+  }
+
+  def unifiedDiff(obtained: String, expected: String): String = {
+    create(obtained, expected).unifiedDiff
+  }
+
+}
