+++ tests/shared/src/test/scala/munit/DiffsSuite.scala
@@ -1,12 +1,10 @@
 package munit
 
-import munit.internal.difflib.Diffs
-
 class DiffsSuite extends FunSuite { self =>
   test("ansi") {
-    val diff1 = Diffs.unifiedDiff("a", "b")
-    val diff2 = Diffs.unifiedDiff("a", "c")
-    val obtained = Diffs.unifiedDiff(diff1, diff2)
+    val diff1 = munit.diff.Diffs.unifiedDiff("a", "b")
+    val diff2 = munit.diff.Diffs.unifiedDiff("a", "c")
+    val obtained = munit.diff.Diffs.unifiedDiff(diff1, diff2)
     // Asserts that a roundtrip of ANSI color processing still produces
     // intuitive results.
     assertNoDiff(
@@ -25,7 +23,7 @@ class DiffsSuite extends FunSuite { self =>
       expected: String
   )(implicit loc: Location): Unit = {
     test(name) {
-      val obtained = Diffs.unifiedDiff(a, b)
+      val obtained = munit.diff.Diffs.unifiedDiff(a, b)
       assertNoDiff(obtained, expected)
     }
   }
